+++ angular/src/lib/modal/index.ts
@@ -1,9 +1,8 @@
 export * from './modal.module';
-export * from './modal.component';
-export * from './modal-body.module';
 export * from './modal-body.component';
-export * from './modal-footer.module';
+export * from './modal-container.component';
 export * from './modal-footer.component';
-export * from './modal-header.module';
 export * from './modal-header.component';
-
+export * from './modal-ref';
+export * from './modal.component';
+export * from './modal.service';