+++ react/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [20.1.3](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/react@20.1.2...@momentum-ui/react@20.1.3) (2019-07-29)
+
+**Note:** Version bump only for package @momentum-ui/react
+
+
+
+
+
+## [20.1.2](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/react@20.1.1...@momentum-ui/react@20.1.2) (2019-07-26)
+
+
+### Bug Fixes
+
+* **List:** fix active/focus context on list ([d33929a](https://github.com/momentum-design/momentum-ui/commit/d33929a))
+
+
+
+
+
 ## [20.1.1](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/react@20.1.0...@momentum-ui/react@20.1.1) (2019-07-09)
 
 **Note:** Version bump only for package @momentum-ui/react